+++ modules/api/controllers/CountriesController.php
@@ -4,9 +4,9 @@
 namespace app\modules\api\controllers;
 
 
-use app\modules\api\common\ApiController;
+use app\modules\api\common\BaseActiveController;
 
-class CountriesController extends ApiController
+class CountriesController extends BaseActiveController
 {
     public $modelClass = 'app\models\activerecord\Countries';